app/requirements.txt
# Core requirements
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.24.0

# File handling
openpyxl>=3.1.0
xlrd>=2.0.1
pyyaml>=6.0
python-dotenv>=1.0.0

# Performance improvements
#pyarrow==12.0.1